Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
cmake==3.31.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dlib==19.24.6
exceptiongroup==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.5
fonttools==4.55.0
h11==0.14.0
idna==3.10
imageio==2.36.0
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
Pillow==9.5.0
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
scikit-build==0.18.1
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tifffile==2024.9.20
tomli==2.1.0
typing_extensions==4.12.2
uvicorn==0.32.1
